+++ SnowboardCourses/src/main/java/ru/tsedrik/dao/LocationDAO.java
@@ -3,6 +3,6 @@ package ru.tsedrik.dao;
 import ru.tsedrik.model.CourseLocation;
 import ru.tsedrik.model.Person;
 
-public interface LocationDAO extends GenericDAO<CourseLocation, Integer> {
+public interface LocationDAO extends GenericDAO<CourseLocation, Long> {
 
 }
