+++ SnowboardCourses/src/main/java/ru/tsedrik/service/GroupService.java
@@ -5,6 +5,6 @@ import ru.tsedrik.model.Group;
 public interface GroupService {
     void addGroup(Group group);
     Group deleteGroup(Group group);
-    Group deleteGroupById(int id);
-    Group getGroupById(int id);
+    Group deleteGroupById(Long id);
+    Group getGroupById(Long id);
 }
